requirements.txt
dj-database-url
psycopg2
django == 5.0.2
djangorestframework== 3.14.0
gunicorn
whitenoise==6.5.0
djangorestframework-simplejwt
django-cors-headers
